+++ Implementacion/src/main/java/Platos/PlatoSimple.java
@@ -4,15 +4,15 @@ import java.math.BigDecimal;
 import java.util.List;
 
 public class PlatoSimple extends Plato {
-    public PlatoSimple(String nombre, Double precio, List<String> ingredientes){
+    public PlatoSimple(String nombre, double precio, List<String> ingredientes){
         super(nombre);
-        this.precio = BigDecimal.valueOf(precio);
+        this.precio = precio;
     }
     List<String> ingredientes; //TODO: ver despues como implementamos esto
-    BigDecimal precio;
+    double precio;
 
     @Override
-    public BigDecimal getPrecio() {
+    public Double getPrecio() {
         return precio;
     }
 }
